EarnableScene base phaser class
