+++ src/components/Play/games/GymRoom.jsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect, useContext } from "react";
 import Phaser from "phaser";
-import { useMoralis } from "react-moralis";
 import { IonPhaser } from "@ion-phaser/react";
 import { GymRoomScene } from "./GymRoomScene";
 import { SpaceStretchScene } from "./SpaceStretchScene";
